+++ sdk/src/main/java/com/twincoders/twinpush/sdk/communications/DefaultRequestLauncher.java
@@ -2,14 +2,13 @@ package com.twincoders.twinpush.sdk.communications;
 
 import android.content.Context;
 import android.os.Handler;
-import android.support.annotation.NonNull;
+import androidx.annotation.NonNull;
 
 import com.android.volley.Request;
 import com.android.volley.RequestQueue;
 import com.android.volley.toolbox.Volley;
 import com.twincoders.twinpush.sdk.logging.Ln;
 
-import java.nio.charset.StandardCharsets;
 import java.util.HashMap;
 import java.util.Map;