+++ sdk/src/main/java/com/twincoders/twinpush/sdk/communications/RESTRequest.java
@@ -15,33 +15,40 @@ import java.util.Map;
 public abstract class RESTRequest extends DefaultRequest {
 
 	List<String> segmentParams = new ArrayList<String>();
-	
+
 	protected String baseURL = null;
 	protected String resourceName = null;
 	protected String methodName = null;
-	
+
 	public RESTRequest() {
 		this.httpMethod = HttpMethod.GET;
 	}
-	
+
 	public void addSegmentParam(String param) {
 		segmentParams.add(param);
 	}
-	
+
 	protected void serializeSegments(List<String> segments, Uri.Builder b) {
 		for (String segment : segments) {
 			b.appendPath(segment);
 		}
 	}
-	
+
 	protected void serializeParams(List<TwinRequestParam> params, Uri.Builder b) {
 		if (this.httpMethod == HttpMethod.GET) {
 			for (TwinRequestParam param : params) {
-				b.appendQueryParameter(param.getKey(), param.getValue().toString());
+				if (param.getParamType() == TwinRequestParam.ParamType.ARRAY) {
+					String paramName = param.getKey() + "[]";
+					for (String value: param.getArrayValue()) {
+						b.appendQueryParameter(paramName, value);
+					}
+				} else {
+					b.appendQueryParameter(param.getKey(), param.getValue().toString());
+				}
 			}
 		}
 	}
-	
+
 	@Override
 	public String getURL() {
 		String url = "";
@@ -55,9 +62,9 @@ public abstract class RESTRequest extends DefaultRequest {
 			}
 			this.serializeSegments(segmentParams, b);
 			this.serializeParams(getParams(), b);
-			url = b.build().toString(); 
+			url = b.build().toString();
 		}
-		
+
 		return url;
 	}
 
@@ -68,32 +75,32 @@ public abstract class RESTRequest extends DefaultRequest {
 	public String getResourceName() {
 		return resourceName;
 	}
-	
+
 	public String getMethodName() {
 		return methodName;
 	}
-	
+
 	protected JSONObject serializeBodyParams(List<TwinRequestParam> params) {
 		JSONObject json = new JSONObject();
 		for (TwinRequestParam param : params) {
 			try {
 				switch (param.getParamType()) {
-				case SIMPLE:
-					if (param.getValue() != null) {
-						if (param.getValue() instanceof Map) {
-							json.put(param.getKey(), new JSONObject((Map<?, ?>) param.getValue()));
-						} else {
-							json.put(param.getKey(), param.getValue());
+					case SIMPLE:
+						if (param.getValue() != null) {
+							if (param.getValue() instanceof Map) {
+								json.put(param.getKey(), new JSONObject((Map<?, ?>) param.getValue()));
+							} else {
+								json.put(param.getKey(), param.getValue());
+							}
 						}
-					}
-					break;
-				case ARRAY:
-					if (param.getArrayValue() != null) {
-						JSONArray array = new JSONArray(param.getArrayValue());
-						json.put(param.getKey(), array);
-					}
-				default:
-					break;
+						break;
+					case ARRAY:
+						if (param.getArrayValue() != null) {
+							JSONArray array = new JSONArray(param.getArrayValue());
+							json.put(param.getKey(), array);
+						}
+					default:
+						break;
 				}
 			} catch (JSONException e) {
 				Ln.e(e, "Error while trying to serialize params");
@@ -101,7 +108,7 @@ public abstract class RESTRequest extends DefaultRequest {
 		}
 		return json;
 	}
-	
+
 	@Override
 	public String getBodyContent() {
 		String body = serializeBodyParams(getParams()).toString();