+++ sdk/src/main/java/com/twincoders/twinpush/sdk/util/LastLocationFinder.java
@@ -27,7 +27,7 @@ import android.location.Criteria;
 import android.location.Location;
 import android.location.LocationListener;
 import android.location.LocationManager;
-import android.support.v4.app.ActivityCompat;
+import androidx.core.app.ActivityCompat;
 
 import com.twincoders.twinpush.sdk.logging.Ln;
 
@@ -65,7 +65,7 @@ public class LastLocationFinder {
         // The calling Activity will likely (or have already) request ongoing
         // updates using the Fine location provider.
         criteria = new Criteria();
-        criteria.setAccuracy(Criteria.ACCURACY_LOW);
+        criteria.setAccuracy(Criteria.ACCURACY_FINE);
 
         // Construct the Pending Intent that will be broadcast by the oneshot
         // location update.